+++ ush/wave_tar.sh
@@ -27,16 +27,16 @@
 
 # 0.a Basic modes of operation
 
-cd "${DATA}"
+cd "${DATA}" || exit 1
 echo "Making TAR FILE"
-
-alertName=$(echo $RUN | tr [a-z] [A-Z])
+export err
 
 # 0.b Check if type set
 
 if [[ "$#" -lt '3' ]]; then
-    echo 'FATAL ERROR: VARIABLES IN ww3_tar.sh NOT SET'
-    exit 1
+    msg='FATAL ERROR: VARIABLES IN ww3_tar.sh NOT SET'
+    err=1
+    err_exit "${msg}"
 else
     ID=$1
     type=$2
@@ -53,27 +53,28 @@ cat << EOF
    Number of files : $3
 EOF
 
-filext=$type
-if [[ "$type" == "ibp" ]]; then
+filext=${type}
+if [[ "${type}" == "ibp" ]]; then
     filext='spec'
 fi
-if [[ "$type" == "ibpbull" ]]; then
+if [[ "${type}" == "ibpbull" ]]; then
     filext='bull'
 fi
-if [[ "$type" == "ibpcbull" ]]; then
+if [[ "${type}" == "ibpcbull" ]]; then
     filext='cbull'
 fi
 
-rm -rf TAR_${filext}_$ID
-mkdir TAR_${filext}_$ID
+rm -rf "TAR_${filext}_${ID}"
+mkdir "TAR_${filext}_${ID}"
 # this directory is used only for error capturing
 
 # 0.c Define directories and the search path.
 #     The tested variables should be exported by the postprocessor script.
 
 if [[ -z "${COMOUT_WAVE_STATION+x}" || -z "${SENDDBN+x}" || -z "${STA_DIR+x}" ]]; then
-    echo 'FATAL ERROR: EXPORTED VARIABLES IN ww3_tar.sh NOT SET'
-    exit 2
+    msg='FATAL ERROR: EXPORTED VARIABLES IN ww3_tar.sh NOT SET'
+    err=2
+    err_exit "${msg}"
 fi
 
 # --------------------------------------------------------------------------- #
@@ -85,24 +86,26 @@ countMAX=5
 tardone='no'
 sleep_interval=10
 
-while [[ "${tardone}" = "no" ]]; do
-
-    nf=$(find . -maxdepth 1 -type f -name "*.$filext" | wc -l)
-    nbm2=$(($nb - 2))
+while [[ "${tardone}" == "no" ]]; do
+    # shellcheck disable=SC2312
+    nf=$(find . -maxdepth 1 -type f -name "*.${filext}" -printf '.' | wc -c)
+    nbm2=$((nb - 2))
     if [[ ${nf} -ge ${nbm2} ]]; then
 
         tar -cf "${ID}.${type}.tar" ./*."${filext}"
         err=$?
 
         if [[ ${err} -ne 0 ]]; then
-            echo 'FATAL ERROR: TAR CREATION FAILED *** '
-            exit 3
+            msg='FATAL ERROR: TAR CREATION FAILED *** '
+            err=3
+            err_exit "${msg}"
         fi
 
         filename="${ID}.${type}.tar"
         if ! wait_for_file "${filename}" "${sleep_interval}" "${countMAX}"; then
-            echo "FATAL ERROR: File ${filename} not found after waiting $((sleep_interval * (countMAX + 1))) secs"
-            exit 3
+            msg="FATAL ERROR: File ${filename} not found after waiting $((sleep_interval * (countMAX + 1))) secs"
+            err=3
+            err_exit "${msg}"
         fi
 
         if [[ -f "${ID}.${type}.tar" ]]; then
@@ -113,8 +116,9 @@ while [[ "${tardone}" = "no" ]]; do
 done
 
 if [[ "${tardone}" == 'no' ]]; then
-    echo 'FATAL ERROR: TAR CREATION FAILED *** '
-    exit 3
+    msg='FATAL ERROR: TAR CREATION FAILED *** '
+    err=3
+    err_exit "${msg}"
 fi
 
 if [[ "${type}" == 'spec' ]]; then
@@ -124,8 +128,9 @@ if [[ "${type}" == 'spec' ]]; then
         err=$?
 
         if [[ ${err} -ne 0 ]]; then
-            echo 'FATAL ERROR: SPECTRAL TAR COMPRESSION FAILED *** '
-            exit 4
+            msg='FATAL ERROR: SPECTRAL TAR COMPRESSION FAILED *** '
+            err=4
+            err_exit
         fi
     fi
 else
@@ -142,24 +147,25 @@ cpfs "${file_name}" "${COMOUT_WAVE_STATION}/."
 err=$?
 
 if [[ ${err} -ne 0 ]]; then
-    echo 'FATAL ERROR: TAR COPY FAILED *** '
-    exit 4
+    msg='FATAL ERROR: TAR COPY FAILED *** '
+    export err=4
+    err_exit "${msg}"
 fi
 
 if [[ "${SENDDBN}" == "YES" ]]; then
     echo "   Alerting TAR file as ${COMOUT_WAVE_STATION}/${file_name}"
-    "${DBNROOT}/bin/dbn_alert MODEL" "${alertName}_WAVE_TAR" "${job}" \
+    "${DBNROOT}/bin/dbn_alert MODEL" "${RUN^^}_WAVE_TAR" "${job}" \
         "${COMOUT_WAVE_STATION}/${file_name}"
 fi
 
 # --------------------------------------------------------------------------- #
 # 4.  Final clean up
 
-cd "${DATA}"
+cd "${DATA}" || exit 1
 
-if [[ ${KEEPDATA:-NO} == "NO" ]]; then
+if [[ "${KEEPDATA:-NO}" == "NO" ]]; then
     set -v
-    rm -rf ${STA_DIR}/${type}
+    rm -rf "${STA_DIR:?}/${type}"
     set +v
 fi
